moved options to specify keys from metadata command to slim-metadata, updated units accordingly, added slim_metadata defaults to cromshellconfig.
