+++ src/cromshell/slim_metadata/command.py
@@ -11,26 +11,38 @@ LOGGER = logging.getLogger(__name__)
 @click.command(name="slim-metadata")
 @click.argument("workflow_id")
 @click.option(
+    "--key",
+    "-k",
+    multiple=True,
+    help="Use keys to get a subset of the metadata for a workflow. Multiple keys "
+    "can be set by add option name before adding key (e.g. '-k id -k status ...')",
+)
+@click.option(
     "--not_expand_subworkflow",
     is_flag=True,
     default=True,
     help="Do not expand subworkflow info in metadata",
 )
+@click.option(
+    "-e",
+    "--exclude_keys",
+    is_flag=True,
+    show_default=True,
+    default=False,
+    help="Toggle to either include or exclude keys that are specified "
+         "by the --keys option or in the cromshell config JSON.",
+)
 @click.pass_obj
-def main(config, workflow_id: str, not_expand_subworkflow: bool):
+def main(
+        config,
+        workflow_id: str,
+        key: list,
+        not_expand_subworkflow: bool,
+        exclude_keys: bool
+):
     """Get a subset of the workflow metadata using default keys."""
 
     LOGGER.info("slim-metadata")
-
-    keys = [
-        "id",
-        "executionStatus",
-        "backendStatus",
-        "status",
-        "callRoot",
-        "subWorkflowMetadata",
-        "subWorkflowId",
-    ]
 
     # Overrides the default cromwell url set in the cromshell config file or
     # command line argument if the workflow id is found in the submission file.
@@ -41,11 +53,19 @@ def main(config, workflow_id: str, not_expand_subworkflow: bool):
     # Check if Cromwell Server Backend works
     http_utils.assert_can_communicate_with_server(config)
 
-    LOGGER.info("Metadata keys set to: %s", keys)
+    # Resolve and get metadata keys from cli, config file, or config default
+    metadata_parameter = resolve_and_return_metadata_keys(
+        cli_key=key,
+        cromshell_config_options=config.cromshell_config_options,
+        config_slim_metadata_default_param=config.SLIM_METADATA_PARAMETERS,
+    )
+
+    key_action = "include" if not exclude_keys else "exclude"
+    LOGGER.info("Metadata keys set to %s: %s", key_action, metadata_parameter)
 
     combined_metadata_parameter = metadata_command.combine_keys_and_flags(
-        list_of_keys=keys,
-        exclude_keys=False,
+        list_of_keys=metadata_parameter,
+        exclude_keys=exclude_keys,
         not_expand_subworkflow=not_expand_subworkflow,
     )
 
@@ -60,3 +80,29 @@ def main(config, workflow_id: str, not_expand_subworkflow: bool):
     io_utils.pretty_print_json(workflow_metadata_json, add_color=True)
 
     return 0
+
+
+def resolve_and_return_metadata_keys(
+    cli_key: list,
+    cromshell_config_options: dict,
+    config_slim_metadata_default_param: list,
+) -> list:
+    """Determines which metadata keys to use from cli, config file, and default
+    parameters, then returns a string of the processed keys ready to be used
+    in an api call. Also returns true if default parameter is being used else false."""
+
+    # If keys is specified in cli then use this first
+    if cli_key:
+        LOGGER.info("Using metadata key(s) from command line options.")
+        return cli_key
+
+    # If metadata_keys is specified in cromshell config file then use it for keys
+    elif "slim_metadata_keys" in cromshell_config_options:
+        LOGGER.info("Setting metadata key(s) from value in config file.")
+        return cromshell_config_options["slim_metadata_keys"]
+
+    # Return the default keys from config module constant
+    else:
+        LOGGER.info("No metadata keys were found in cli or config file.")
+        # The default metadata key needs to
+        return config_slim_metadata_default_param
