+++ src/cromshell/utilities/cromshellconfig.py
@@ -11,7 +11,16 @@ LOGGER = logging.getLogger(__name__)
 """Setup Cromshell config details. Intended to be used as a singleton"""
 
 # Set Cromshell Configuration Default Values
-METADATA_PARAMETERS = {'keys': ["submittedFiles"], 'includeKeys': False}
+METADATA_PARAMETERS = ["submittedFiles"]
+SLIM_METADATA_PARAMETERS = [
+                            "id",
+                            "executionStatus",
+                            "backendStatus",
+                            "status",
+                            "callRoot",
+                            "subWorkflowMetadata",
+                            "subWorkflowId",
+                           ]
 API_STRING = "/api/workflows/v1"
 # Concatenate the cromwell url, api string, and workflow ID. Set in subcommand.
 cromwell_api_workflow_id = None
