+++ tests/unit/test_metadata.py
@@ -76,53 +76,3 @@ class TestMetadata:
             )
             == test_keys_string_out
         )
-
-    @pytest.mark.parametrize(
-        "test_keys, "
-        "test_cromshell_config_options, "
-        "test_config_metadata_param, "
-        "out_str, "
-        "use_defaults",
-        [
-            # Cli keys should be used if provided
-            (
-                ["id"],
-                {"metadata_keys": "key"},
-                {'keys': ["submittedFiles"], 'includeKeys': False},
-                ["id"],
-                False,
-            ),
-            # Config keys should be used if cli empty
-            (
-                [],
-                {"metadata_keys": ["key"]},
-                {'keys': ["submittedFiles"], 'includeKeys': False},
-                ["key"],
-                False,
-            ),
-            # Metadata should be used by default if no keys found in cli or config
-            (
-                [],
-                {"cromwell": ["stuff"]},
-                {'keys': ["submittedFiles"], 'includeKeys': False},
-                ["submittedFiles"],
-                True,
-            ),
-        ],
-    )
-    def test_resolve_and_return_metadata_keys(
-        self,
-        test_keys,
-        test_cromshell_config_options,
-        test_config_metadata_param,
-        out_str,
-        use_defaults,
-    ):
-        assert (
-            metadata_command.resolve_and_return_metadata_keys(
-                cli_key=test_keys,
-                cromshell_config_options=test_cromshell_config_options,
-                config_metadata_param=test_config_metadata_param,
-            )
-            == out_str, use_defaults
-        )
